requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain-groq
langchain
jinja2
pydantic